backend_python/requirements.txt
# FastAPI Backend Dependencies
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM and AI libraries
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Data processing
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# CORS and middleware
python-cors==1.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.9.1
flake8==6.1.0
mypy==1.6.1
